sesion21/requirements.txt
django
djangorestframework
pymysql
cryptography
django-cors-headers